Add create backup function and hide backup dir in gitignore
